zeta2: Split up crate into modules
